requirements.txt
django==5.2.6
djangorestframework==3.16.1
django-filter==25.1
django-cors-headers==4.9.0
django-celery-results==2.6.0
google-cloud-vision==3.10.2
google-cloud-storage==2.18.0
python-dotenv==1.1.1
celery==5.5.3
redis==6.4.0
gevent==25.9.1
Pillow==11.3.0
psycopg2-binary==2.9.10
django-taggit==6.1.0